codebase-genius/requirements.txt
jaclang>=0.7.0
gitpython>=3.1.40
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0